# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas Platform API
  version: 0.2.0
  description: Platform operations — orgs, onboarding, webhooks, events, reports, exports, and cron jobs
  contact:
    name: AtlasPayments Team
    url: https://github.com/shivemind/atlaspayments
servers:
  - url: http://localhost:3000
    description: Local development server. Override via Postman {{baseUrl}} environment variable.
...
paths:
...
  /api/internal/underwriting/cases/{id}:
    get:
      summary: Get underwriting case
      description: Returns a single underwriting case by ID.
      x-platform:
        kind: resource
        entity: UnderwritingCase
        affectsLedger: false
        workflowRefs:
          - onboarding_full
      operationId: getUnderwritingCase
      tags:
        - Underwriting
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Underwriting case found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/UnderwritingCase"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Case not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update underwriting case
      description: Updates assignment, risk score, notes, or status of an underwriting case.
      x-platform:
        kind: resource
        entity: UnderwritingCase
        emitsEvents:
          - underwriting.case.updated
        affectsLedger: false
      operationId: updateUnderwritingCase
      tags:
        - Underwriting
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                assigned_to:
                  type: string
                risk_score:
                  type: integer
                  minimum: 0
                  maximum: 1000
                notes:
                  type: string
                status:
                  type: string
      responses:
        "200":
          description: Underwriting case updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/UnderwritingCase"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Case not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/internal/reports:
    get:
      summary: List all reports (admin)
      description: Returns a paginated list of all reports across merchants.
      x-platform:
        kind: resource
        entity: Report
        affectsLedger: false
      operationId: adminListReports
      tags:
        - Reports
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - in: query
          name: merchant_id
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum:
              - queued
              - processing
              - completed
              - failed
      responses:
        "200":
          description: Paginated list of reports.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Report"
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/internal/orgs/{orgId}/teams/{teamId}:
    get:
      summary: Get team
      description: Returns a single team by ID.
      x-platform:
        kind: resource
        entity: Team
        affectsLedger: false
      operationId: getTeam
      tags:
        - Teams
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - *a3
        - &a6
          in: path
          name: teamId
          required: true
          description: Team ID.
          schema:
            type: string
      responses:
        "200":
          description: Team details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "401": *a1
        "404": *a5
    patch:
      summary: Update team
      description: Updates fields on a team.
      x-platform:
        kind: resource
        entity: Team
        affectsLedger: false
      operationId: updateTeam
      tags:
        - Teams
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - *a3
        - *a6
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                metadata:
                  type: object
      responses:
        "200":
          description: Team updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "400": *a4
        "401": *a1
        "404": *a5
    delete:
      summary: Delete team
      description: Deletes a team by ID.
      x-platform:
        kind: resource
        entity: Team
        affectsLedger: false
      operationId: deleteTeam
      tags:
        - Teams
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - *a3
        - *a6
      responses:
        "200":
          description: Team deleted.
          content:
            application/json:
              schema:
                type: object
                required:
                  - deleted
                  - id
                properties:
                  deleted:
                    type: boolean
                  id:
                    type: string
        "401": *a1
        "404": *a5
  /api/internal/orgs/{orgId}/teams/{teamId}/members:
    get:
      summary: List team members
      description: Returns a paginated list of members for a team.
      x-platform:
        kind: resource
        entity: TeamMember
        affectsLedger: false
      operationId: listTeamMembers
      tags:
        - Teams
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - *a3
        - *a6
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of team members.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401": *a1
        "403": *a2
        "404": *a5
    post:
      summary: Add team member
      description: Adds a member to a team.
      x-platform:
        kind: resource
        entity: TeamMember
        affectsLedger: false
      operationId: addTeamMember
      tags:
        - Teams
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - *a3
        - *a6
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
              properties:
                user_id:
                  type: string
                role:
                  type: string
      responses:
        "201":
          description: Member added to team.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "400": *a4
        "401": *a1
        "404": *a5
...
components:
  securitySchemes:
    MerchantKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for merchant-scoped operations. Pass as `Authorization: Bearer {{merchantApiKey}}`.
    PlatformAdminKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for platform admin operations. Pass as `Authorization: Bearer {{platformAdminKey}}`.
  schemas:
    ErrorResponse:
      description: Standard error response envelope returned for 4xx/5xx responses.
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code.
            message:
              type: string
              description: Human-readable error description.
    PaginationMeta:
      description: Pagination metadata included in list responses.
      type: object
      required:
        - page
        - pageSize
        - total
        - hasMore
      properties:
        page:
          type: integer
          description: Current page number (1-indexed).
        pageSize:
          type: integer
          description: Number of results per page.
        total:
          type: integer
          description: Total number of matching records.
        hasMore:
          type: boolean
          description: Whether additional pages exist.
...
    UnderwritingCase:
      description: Underwriting review case linked to an onboarding profile.
      type: object
      properties:
        id:
          type: string
        onboarding_profile_id:
          type: string
        assigned_to:
          type: string
          nullable: true
        risk_score:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Report:
      description: An on-demand report generated for a merchant.
      type: object
      properties:
        id:
          type: string
        merchant_id:
          type: string
        type:
          type: string
        name:
          type: string
        status:
          type: string
          enum:
            - queued
            - processing
            - completed
            - failed
        parameters:
          type: object
          nullable: true
        result_url:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  parameters:
...
    PageParam:
      in: query
      name: page
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      in: query
      name: pageSize
      description: Number of results per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    PathId:
      in: path
      name: id
      required: true
      description: Resource identifier.
      schema:
        type: string
...